GH-31 Allow selection of GitProviderStrategy via configuration property
The configuration property is received via the ApplicationContext of the running test (this should be fine, since a relevant test should always be an @ApplicationModuleTest and thus, by meta-annotation, have a Spring ApplicationContext).

The strategy instance is then looked up via a ServiceLoader. This is meant as an SPI/extension point to allow custom implementations in the future.